dump: In modemRx header add att and rxpow columns
In the modem Rx status dump, each line ends with a list of all
possible paths (0..N-1) attenuations and actually received powers.
But as the list has a variable length there was no headings at all.
Now we print a heading at least for the 1st column so people can
try to guess the meaning of the remainding columns.

Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>