optimize rx_search
Optimize how the Rx search is done:
Before, devices in the Rx_search state were constantly looking
for a matching Tx.
Instead each time a transmission is started we go over the
devices currently in rx_search state and, if they could synchronize to
this transmission, we change their state to a new "rx_found" state.
This reduces significantly the number of function queue operations.

This commit should not change in any way the external behaviour of the phy.

Signed-off-by: Paul Quevedo <paulquevedo@jdrfelectromag.com>
Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>