optimization: Faster simulations by reducing number of rx_search operations
This commit changes the phy to track the devices currently
searching for a suitable transmission. Each time a transmission is
executed the devices in the search list attempt to synchronize. This
reduces the number of redundant function queue operations that occur as
a result of executing an rx search operation at every time step.

Signed-off-by: Paul Quevedo <paulquevedo@jdrfelectromag.com>