Add cmd line arg to pass failed pkts to device
Add a command line option to select if the Phy should send the
original packet to the device attempting to receive it, even if the
packet has a CRC error or reception is aborted midway. By default
this is not the case.

This change enables to simulate the behavior of some radios more
accurately, if the device handles it correctly. However, the data is
delivered as it was sent, i.e., the delivered packet does not have bit
errors.

Signed-off-by: Wolfgang Puffitsch <wopu@demant.com>