Add v2 API support
Both add handling of the new v2 API procedures,
and the internal logic that allows handling many of
the new API functionality.

Internally both the new and old API are handled by the
same logic.
But note that the Phy remains fully compatible with the old API,
and that its behaviour is mostly unchanged when using old API
procedures.

Note that not all new API provided functionality is supported yet.
At this point the following are missing:
 * Prelocked_tx: piggybacking receptions
 * Forced_packet_duration: receiver incorrectly detecting the packet lenght
 * Support for != coding_rate in Tx and Rx
 * Support for immediate bit error masks

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>