dump_postprocess: pcap & pcapng: Handle better truncated dumps
When feeding truncated dumps to these scripts, they would
throw an exception.
Instead let's handle these cases better.
If a line is choped mid way badly, we just consider the file
ended, and ignore the line.
If a line has everything but a truncated packet, we
write as much of the packet out as we have (wireshark will just
see the packet as malformed).

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>