main: Add support for prelocked_tx
Add support for prelocked_tx in the main state machine.

Piggybacking in this change: The Rx_CCA_meas priority was lowered,
to ensure it measures with all Tx in that us already active.

*EXACT EQUALITY CHECK BREAK*
Note: As the Tx_start priority has been raised
the order of events evaluation has changed. This does not
result in a change which makes the phy behave conceptually
differently. But it can cause in some cases differences
in the exact execution result compared with previous versions:
* An possible Rx match will be evaluated a bit later (in order) than
  before and therefore the random draws can differ from that point on
  leading to bit/sync errors in different places
* The Rx response may carry different information about not sync'ed
  Tx's (When a Tx starts to be locked its info is already copied to
  the Rx response). So you may see in dumps or in devices
  differences in the responses content for failed receptions.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>