Bit error calc: When packet is undecodable BER should be 50%
The maximum BER should be 50%, not 100% as we had before.
So let's set it so when the packet has been lost of cannot be
decoded otherwise.

Note this change will produce a change of behaviour compared to
old simulations. Therefore running the simulator in check mode
across this commits will produce different results.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>