modem_if: Pass also coding rate to digital performance functions
To allow the modem digital performance models to account for
the difference due to different coding rates, let's pass
it also this value.
Note: The modem Rx dump is not yet dumping this new value.

NOTE: This is an API change to the modem models IF.
All 2G4 modem models will need to be updated to this new API.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>