feat: support for coded phy packet and power level
Unlike the BLE 1M and BLE 2M packets, CODED PHY packets contain
two entries in results/<sim_id>/*.Tx.csv files (one for FEC1 and a
second for FEC2). This modification takes this into account when
creating a BLE entry for Wireshark.
This modification also adds the power of transmissions for Wireshark.

Signed-off-by: Louka Yerly <louka.yerly@hes-so.ch>